requirements.txt
arrow==0.7.0
click==6.2
Django==1.9.5
gevent==1.0.2
greenlet==0.4.9
gunicorn==19.4.1
psycopg2==2.6.1
python-dateutil==2.4.2
xpinyin==0.5.4
#redis==2.10.5
requests==2.9.0
rq==0.5.6
six==1.10.0
wheel==0.26.0
openpyxl==2.3.5
django-simple-captcha==0.5.1
pyyaml==3.11
pytz
#djangorestframework==3.6.1
#django-filter==1.0.1
pycrypto==2.6.1
python-jenkins==0.4.15
Fabric==1.14.0
ansible==1.9.6
Jinja2==2.10
paramiko==2.4.0
xlrd==1.1.0
MySQL-python==1.2.5
PyMySQL==0.8.0